python/requirements.txt
easyocr
pyrealsense2
opencv-python
paho-mqtt